# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News project API",
        "version": "1"
    },
    "paths": {
...
        "/admin/users/{id}/favourites": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Retrieve a user's favourites",
                "description": "Retrieve a paginated list of a user's favourite items, filtered by user ID and ordered by creation date in descending order.",
                "operationId": "f968d130fbfc09a7e855283659ea021b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The UUID of the user whose favourites to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "favourites": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserFavourite"
                                            }
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "perPage": {
                                            "type": "integer",
                                            "example": 7
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/favourites": {
            "get": {
                "tags": [
                    "User Favourites"
                ],
                "summary": "Get list of user's favourite news",
                "description": "Returns list of user's favourite news",
                "operationId": "getFavourites",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserFavourite"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Favourites"
                ],
                "summary": "Add a news to user's favourites",
                "description": "Add a news to user's favourites",
                "operationId": "addFavourite",
                "requestBody": {
                    "description": "Request body for adding a favourite",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserFavouriteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favourite successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserFavourite"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favourites/{id}": {
            "delete": {
                "tags": [
                    "User Favourites"
                ],
                "summary": "Delete a favourite item",
                "description": "Deletes a favourite item with the specified ID.",
                "operationId": "deleteFavourite",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the favourite item to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a110d11b-176a-4813-8c31-a69f9f50f7fc"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Favourite item successfully deleted"
                    },
                    "404": {
                        "description": "Favourite item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserFavouriteInput": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "My favourite news"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/news/123"
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is my favourite news"
                    },
                    "imageUrl": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "userId": {
                        "type": "string",
                        "example": "93504d6a-e476-48d8-be86-df361cf2c4bb"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "country": {
                        "description": "User country",
                        "type": "string"
                    },
                    "language": {
                        "description": "User language",
                        "type": "string"
                    },
                    "category": {
                        "description": "User category",
                        "type": "string"
                    },
                    "role": {
                        "description": "User role",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Date and time of user creation",
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Date and time of user update",
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "favourites": {
                        "description": "User's favourite news articles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserFavourite"
                        }
                    },
                    "comments": {
                        "description": "User's comments on news articles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserComment"
                        }
                    },
                    "logs": {
                        "description": "User's logs of activity on the news portal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserLog"
                        }
                    }
                },
                "type": "object"
            },
            "UserComment": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "readOnly": true,
                        "example": "f6ebff50-aa5d-11eb-aebf-0242ac130002"
                    },
                    "userUuid": {
                        "type": "string",
                        "example": "f6ebff50-aa5d-11eb-aebf-0242ac130002"
                    },
                    "url": {
                        "type": "text",
                        "example": "https://example.com/article"
                    },
                    "commentText": {
                        "type": "text",
                        "example": "This is a comment"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-05-08T12:30:00Z"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2023-05-08T12:30:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserFavourite": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": "true"
                    },
                    "title": {
                        "type": "string"
                    },
                    "url": {
                        "type": "text"
                    },
                    "author": {
                        "type": "string"
                    },
                    "description": {
                        "type": "text"
                    },
                    "imageUrl": {
                        "type": "text"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserLog": {
                "title": "UserLog",
                "description": "User log model",
                "properties": {
                    "id": {
                        "description": "The unique uuid for the log entry",
                        "type": "string",
                        "format": "uuid"
                    },
                    "action": {
                        "description": "The action performed by the user",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the action performed by the user",
                        "type": "text"
                    },
                    "userId": {
                        "description": "The UUID of the user who performed the action",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time when the log entry was created",
                        "type": "timestamp",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the log entry was last updated",
                        "type": "timestamp",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}